v4.0 (Komekko), 22.01.2021

+++ Cosette.Tuner.Common/Requests/GenerationDataRequest.cs
@@ -6,7 +6,7 @@ namespace Cosette.Tuner.Common.Requests
     {
         public int TestId { get; set; }
         public double ElapsedTime { get; set; }
-        public int BestFitness { get; set; }
+        public double BestFitness { get; set; }
 
         public List<GeneDataRequest> BestChromosomeGenes { get; set; }
     }
